+++ modulenetworkinteractionapp/processrequest/sendMsgPingPong.go
@@ -8,7 +8,7 @@ import (
 )
 
 //SendMsgPingPong отправить сообщение типа Ping
-func SendMsgPingPong(msgType string, countProsessFilter int) ([]byte, error) {
+func SendMsgPingPong(msgType string, countProsessFilter int8) ([]byte, error) {
 	fmt.Println("для клиента, отправить запрос PING")
 
 	msgPing := configure.MsgTypePingPong{
